implement login methods in API's
the login method for the fake and real API classes was implemented.

Since user information is needed in many parts of the app, it will be better to push the current user onto a stream that can be accessed by all pages in app. Because of this, LoginResponse was changed slightly so that is just returns a bool and an error (may be null).