Implement missing api models
Quiz class needed to be implemented since the next step of my design is to implement the home page which has quizzes.

Question class to have an additional quizId field to identify which quiz it belongs to.

User and Progress classes were added to represent user information.

Sample user and quiz documents were added.

Sample question document was changed.