Implement initial QuestionsViewModel
Questions view retrieves the questions for its quiz on initialization.

It has a method to handle selecting a choice, which has not been fully implemented yet, but should navigate in the event of a correct selection, and show a snackbar otherwise.

The method setQuestion tries to choose a random question from the list of questions. If a random question is successfully chosen, it will be set as the current question and the View will rebuild. In the event that a random question could not be chosen, the method will check for quiz completion, changing the view if it is completed, or navigate to an error page in all other cases.