Refactor QuestionsViewModel
This still does not work.

The first iteration of [QuestionsViewModel] pushed a new route each time a new question was needed, which looked silly because you could see a glimpse of the previous page after each question.

The second iteration has fields [currentQuestion] and [nextQuestion] which were used by [QuestionsView] to build a stack of two pages. The idea was to animate a transition using the stack, however, this did not work because the children rebuilt before the animation had time to complete.

For future development, it would be good to try a nested navigator structure, which would handle the downfalls of the first iteration while being much simpler and more effective than the second iteration.