Change [QuizView] hero tap callback
In [QuizView], tapping the picture popped off the current route, which was confusing for users. Due to this, the [onHeroTapped] method was changed to be an empty function. The method was not removed and replaced with an empty anonymous closure so that it is easier to refactor in the future.