Attempt to show nested snackbar
The snackbars that are displayed by [ScaffoldService] are displayed at the root level. This means that if you select an incorrect choice, then move to a different page (such as exiting or finishing the quiz, or going to a new question), the snackbar still persists. The changes here do not fix the issue, but should be on the right track.