Remodel fake data interactions
Since a [user] field was added to [LoginResponse], the return value from [FakeApi] no longer authenticates successfully. This was temporarily changed by adding a conditional which authenticates all fake data returns. It is preferable to look for a new way to do this, since it does not allow testing of login errors.