fix: Image wouldn't render on cold load
On a cold boot, the [FittedBox] raised an error that there was no child size available for building. This probably happened because the image took too long to load. This can be fixed by wrapping it in [SizedBox.expand], which gives [FittedBox] tight constraints so it doesn't need a child size to build.