Change data parsing in constructor
This raised a /lot/ of issues. The issue was that Firestore returns [Map<String, List<dynamic>>] which cannot be easily cast into [Map<String, List<String>>].

This was solved by taking in [Map<String, dynamic>] in the [Progress] constructor, and then parsing the data inside the constructor using [List.from]. This definitely taught me to handle data parsing at the lowest level possible so that you don't have to chase return values around for hours.